Cast to size_t before doing comparison (#420)

+++ rmw_connext_cpp/src/rmw_take.cpp
@@ -249,7 +249,7 @@ _take_sequence(
     return RMW_RET_ERROR;
   }
 
-  if (count > (std::numeric_limits<DDS_Long>::max)()) {
+  if (count > static_cast<size_t>((std::numeric_limits<DDS_Long>::max)())) {
     RMW_SET_ERROR_MSG_WITH_FORMAT_STRING(
       "cannot take %ld samples at once, limit is %d",
       count, (std::numeric_limits<DDS_Long>::max)());
